Added two workflows and .editorconfig
The workflows are:
    - dependabot
    - check

Dependabot just checks that dependencies are safe and upgraded if needed.

Check adds a bunch of things I feel are needed.

Note: Both of these workflows come from: https://github.com/jonhoo/rust-ci-conf

An .EditorConfig is provide as it is very important to have the same style
everywhere regardless of editor.