Add algebra symbols
Fixes #309 with `tri.l.eq`.

I made a distinction between decorative `triangle`s and mathematical `tri`s; I feel this warrants some consideration and discussion before the final merge. Note also the terminology for `times` (not to be confused with the various flavors of `product`) and the semidirect product symbols, `times.l` and `times.r`.