fix: allow access to label of sequence
Previously, the following file would fail to compile

```typ
 #show: it => {
   let sequence = it.children.at(0)
   repr(sequence)
   sequence.label
   //       ^^^^^~~~ error: content does not contain field "label"
 }
 #[
   = Heading
   text
 ] <label>
```

however, both the `repr` and the IDE autocomplete say that the field
"label" does indeed exist in the sequence. This commit fixes the issue
by falling back to regular field access on `Content` when it is a
sequence and the user isn't trying to access the "children" field.