Add Chinese numbering
It is worth noting that the case system in Rust's standard library does not
support Chinese characters since most Chinese scripts and punctuation lack
case. Therefore, the `is_lowercase()` and `is_uppercase()` functions will both
return false for Chinese characters.

In Chinese, there are two character sets used to represent numbers. One is for
everyday writing, and the other is for commercial use. The commercial set was
created to prevent forgeries.

Although these two character sets are not technically different cases, the
"chinese_number" crate uses uppercase to refer to the commercial numbering and
lowercase to refer to the normal numbering.