Fix grid-cell misalignment bug (#963)
Skip the first region in a cell in a grid row if it's empty before
calculating the size.

Having a table span multiple pages can cause an alignment bug in the
table itself. If the first region in a cell in a grid row is empty, all
other cells in this row will skip the first region. A misalignment bug
can occur, since the calculation of all region sizes happen before the
skip. The overall size allocated for the content of a cell with multiple
regions and content in the first region thus is too little.

Fixes #963