File access bugfixes
Fixed a bug in the cross-origin access checker, that prevented all
accesses.
The checker will now verify for legality, and only allow accesses
whose closest root are their mod's root
(that is, will only allow write access if the write root dir. (dest)
is closer than the read root directory).
It will not try to prevent accesses out of any root.
We have also made the source file's path canonical,
though we may revert this (and allow COA for .typ files) in the future.
(note here)