Classify `dots.c`, `dots.down`, `dots.up` and `dots.v` as normal
Fixes #713.

They were previously classified by Unicode as relations, which caused a '+' sign after one of them to be treated as a unary operator rather than a binary operator. This PR overrides the Unicode classification and classifies them as normal instead, which causes a '+' sign after one of them to be properly treated as a binary operator.