Disallow block comments before markers
Correctly parsing headings, lists, enums, and terms can be difficult
for both humans and computers, if a comment is placed in front.

An example that is difficult to parse for humans would be:

    - foo
      - bar
        /* comment
    */- baz

See #762 for more examples.

However, this change is primarily motivated by the difficulty in parsing
such expressions using
[tree-sitter](https://github.com/frozolotl/tree-sitter-typst).
Producing useful and consistent syntax trees is quite difficult due to
limitations of the tool.

I don't think this change will cause much breakage. Comments before
markers seem niche and not very useful to me.