Make Oklab and Oklch chroma ratios relative to 0.4
In CSS, `oklab(50% 100% 50%)` produces the same color as `oklab(50% 0.4 0.2)`.
The same is true with oklch's chroma parameter: `oklch(50% 100% 0deg) = oklch(50% 0.4 0deg)`.
This commit changes Typst's behavior to be in line with CSS's, which IMO is more intuitive overall.
It also adds increased compatibility with any color pickers that output CSS oklch calls.