Only clear Typst's output, not the entire terminal.
Prior to this PR, the entire terminal was cleared when running `typst watch`.
This is somewhat disruptive to the user experience.
This PR changes that behavior and only clears what Typst has written.

Because to my knowledge there isn't an easy to use ANSI escape sequence for this, I centralized all terminal output in one struct: `TermOut`.
It records the number of lines written and can use that to clear that amount if requested.
This also has the nice benefit of removing certain escape sequence machinery from unrelated functions.

I don't particularly like having to pass this `TermOut` struct everywhere, but it's the best solution I could come up with.
Neither do I like it relying on a `Rc<RefCell<_>>` type thing, but it appears to be necessary due to the way `typst::World` works.

Overall though I am happy with it and I think it is totally usable.