Use alternate screen to not clear the entire terminal
Due to popular demand, I changed the behavior as follows:
By default, fullscreen mode is used. It differs from the implementation on main in that it uses an alternate screen.
That means, once you exit `typst watch`, the prior terminal state reappears.
If enabled via --no-fullscreen, the terminal is not cleared and no alternative screen appears.

What I don't like about this implementation:
- It is more complex because it requires signal handling.
- In fullscreen mode, the last compilation output disappears on exit.
  I might change that in the near future.