Remove `a` suffix from Rgba and LinearRgba
It doesn't really matter, but it has bugged me a bit for a week.
I think the types are a bit inconsistent with the `a` "alpha" suffix: some types have it, others don't.
This PR removes the types' `a` suffix from Rgba and LinearRgba.
I'm not totally sure if it isn't better to add the `a` suffix to all the colors besides luma, but this is what I decided on for now.