Implement barebones JPEG XL support
This commit implements JPEG XL support that appears to work correctly.
However, I wrote this code mostly as a PoC.
That means, the code isn't _great_ and CMYK images can't be imported yet.
I also didn't check if `jxl-oxide` was the best library to use and how complete it actually is.