Make `color.components`' alpha parameter named
This is a rather small change, but until the parameter rework is a thing, an improvement in readability still.
It changes the `alpha` parameter of `color.components` to be named.
Compare these two lines for example:
```typ
#let (r, g, b) = col.components(false)
#let (r, g, b) = col.components(alpha: false)
```
I think in the latter it is more obvious what the argument does.
It also is a bit more consistent with how user-defined functions work, where you can't _really_ specify whether a positional argument should be optional.

An issue with this change, however, is that it does break some code making use of the optional parameter.