initial support for explicit vline customization
On split_vlines, we always go through every row in the current column,
so we in particular we go through every vline in this column going
through the current row and fold their strokes with the current
automatic line's stroke and with the strokes of the cells around the
vline.

A similar procedure should be done for horizontal lines.