Change default highlight color to be transparent
The prior color doesn't look good when the background color is changed, or there is a background image.
A transparent fill improves that.
On a white background, the new color is very slightly less chromatic, but I believe the difference is low enough not to matter (it still looks good).