Vendor OpenSSL on Linux
On Linux, Typst CLI indirectly depends OpenSSL (via `native-tls`) for network access. `native-tls` defaults to dynamic linking. This broke the cross compilation in the release workflow (#3418).

I really wanted to do the dynamic way (because of security patches), but it just didn't work. The solutions in the cross wiki for OpenSSL all assume glibc rather than musl targets (which we currently don't use). We could switch back to glibc  and build on an older OS (to support older distributions), but even after successfully cross-compiling Typst to `aarch64-unknown-linux-gnu`, I got startup errors like `libssl.so.1.1: cannot open shared object file: No such file or directory`.

Vendoring isn't a great solution (because of security patches), but it is in line with providing the most portable binary (we already use musl). Note that this adds a build-time dependency on `perl` and `make` on Linux, but a C toolchain is already required.

Windows and macOS are not affected by thus and thus the openssl vendoring is guarded by `cfg`s.

Fixes #3418