Implement under/over parens/shells, refactor code
I decided to use the word `shell` instead of turtle or tortoise because I
think it is a bit clearer, the word reads nicer (“overshell” vs “overturtle”),
and there is no associated discussion (they are called “tortoise shells” in
Unicode):

> tortoises and turtles are not the same thing!
~ @Enivex

There will be a sibling PR to rename the symbols too.

Additionally, I refactored some code to make it more readable.
I also fixed a bug introduced in #3262 where the wrong script style was being used.

Closes #3856.