Better handle large numbers from JSON
When deserializing JSON files, a `u64` may be
produced by serde, that will then be converted
to a `i64` when using it as a Typst `Value`.
However, some number that can fit in a `u64`
are too big for a `i64`.

So far, Typst overflowed when doing the
conversion, leading to a negative value where a
large positive value was expected. With this
change, it will now use a floating point number
internally to represent this integer. This comes
at the cost of losing precision, but the value
will be much closer to the expected one.

Other fixes to this issue were:
- a saturating cast. It was not satisfying in my
  opinion because the value you would get would be
  very different from the one you would expect,
  making it useless.
- using a string instead of an actual number. With
  that solution, you cannot easily do math on your
  data. It would also be confusing to have a
  number become of string type when imported in
  Typst.
- report an error. I considered that it was better
  to have imprecise data than being completely
  blocked because of what the JSON file contains.
- adding a parameter to the `json` function to
  tell whether or large numbers should result in
  an error, or in imprecisions. In my opinion,
  this option added quite a lot of complexity for
  a use case that most people will never encounter.