Fix #3700: Deformed stroke for some length values.
The stroke was getting deformed for some length values. This was due to
how the curved corners of the stroke were calculated; since it uses the
quadratic formula to calculate a certain variable, when its discriminant
was negative (aka, an imaginary number), the variable would be set to
NaN, resulting in not calculating the corner correctly. The fix was
simply to set the discriminant to 0, when it was negative (i.e., only
the real part of the complex number), solving the issue.