Do layout short-circuit in flow instead of realization
When layouting just a single element capable of LayoutMultiple, we want to keep expansion instead of disabling it for children. That way, it can be passed through and e.g. bottom alignment within nested containers works as expected. Previously this was done in realization, which made it a bit tricky to work with.