Use texts of the first occurrences for `/ToUnicode` CMap
Resolves #4582
…just by deleting `improve_glyph_sets`!

There are two sources of information for `/ToUnicode`: the `glyph_set` recorded while `write_text`, and `cmap` tables of the font.
`improve_glyph_sets` leverages the font. It was refactored into a function in #4154, but the real code even predates ed6550f (2 years ago).

`improve_glyph_sets` was necessary before ad34763, when a `glyph_set` was a list of glyphs, and we had to search the font (again) for their texts. (Each glyph represents a text, which is a Unicode code point or a sequence of code points (e.g. ligature).)

In ad34763, the `glyph_set` is refactored to a map from glyphs to texts.
Now we have enough information for `/ToUnicode` CMap—no need to search the font.

If the glyph…
- …represents a single character…
    - …and is mapped from only one code point:
        No change.
    - …and is shared by multiple code points (e.g. CJK unified/compatibility):
        `/ToUnicode` changes from the largest code points to the first occurrence, and fixes #4582.
- …represents a sequence of characters (e.g. ligature)…
    - …and they are also encoded as a single code point for compatibility (e.g “fi”/ﬁ):
        `/ToUnicode` changes from a single compatibility code point (ﬁ) to the sequence (fi).
        The behaviour in PDF viewers usually does not change.
    - …and is not encoded in Unicode (e.g. “Th” in Linux Libertine):
        No change.