Add missing under/over-delimiters
Adds functions `underparen` & `overparen` and `undershell` & `overshell`, for parentheses and tortoise shell brackets respectively, to Under/Over in Math.

Fixes incorrect style for an annotation above a delimiter in `layout_underoverspreader` (introduced in #3262).

Fixes typo in `LayoutMath` implementation for `UnderbracketElem` (added in #3096).

Refactors some code in `underover.rs` to make it a bit more readable, taking on comments from #3869.

Closes #3856.