Move paragraph widow and orphan prevention into flow
Rather than just stitching together the frames, we now make the decision to prevent a widow or orphan in the flow. This allows us to take the region size into account.

As a result, if there is simply not enough space, we don't have lines sticking together and potentially skipping a region for no reason.