Line::{stretchability, shrinkability}: return +0.0 when empty
rustc 2024-08-22 changed <f64 as Sum>::sum to return -0.0 for an
empty iterator. When -0.0 is passed as the stretchability or
shrinkability parameter of linebreak::raw_ratio, the ratio value
becomes negative infinity and does not hit the `ratio > 1.0`
check. This causes -inf to be returned as the line ratio from
ratio_and_cost, leading to the `active` variable being incremented
too early in `linebreak_optimized_bounded` and the entire line
breaking process being skipped.