Simplify parsing headings. Not fully sure this is desired.
This is a pretty nice change. The difference from before is that we now wrap
whitespace between the end of a heading and a following label as part of the
heading's content. However, that space is always a trailing space, so including
it shouldn't change the output.

Two tests did rely on this quirk, but I updated them.

The original behavior was introduced in this commit. I'm not fully sure why it
was necessary, as the usize::MAX argument to `markup()` was present at that time.

https://github.com/typst/typst/commit/529d3e10c6b4d973e88b6c295eb22a45ea426e42#diff-852b089381f8af01f5470178d6139c92fc3572536b3c348c7171e03a7cc366f2L161