Change type repr to short name
Some types have a long name (for diagnostics) and a short name (for the identifier itself). For example, the type is spelled `int`, but in error messages it says "expected integer".

Currently, the `repr` uses the long name, but that's rather confusing since the repr output is typically (not always) close to the code that produces the value. Thus, this PR changes the type repr to the short name.

Converting a type to a string (`str(ty)`) will continue to output the long name. Changing that would be a breaking change. Also, it's mostly used for error messages and that's what the long name is for.