Fix param autocompletion false positive
With the cursor _after_ a parameter list (`#text(font:)|`), we would still emit completions, they just wouldn't typically show up in an editor because the emitted cursor position was after the colon and no completion started with `)`. This let the autocompletion fonts test pass even though the indices are wrong.