Add -e flag to winget installation method
The `-e` (`--exact`) flag keeps winget from installing another package that was fuzzily matched. This is not *necessary* here, as the provided `--id` should match only Typst and at the moment, `winget search typst` only returns Typst itself and tinymist anyway, but it is another level of security, which I personally think should be recommended to users when installing.