Ignore CSL's `display="left-margin"` in block-level containers
The CSL `display="left-margin"` attribute is meant for top-level use. The CSL spec isn't clear what should happen when it is used in a container, but not a single style from https://github.com/citation-style-language/styles/ uses it in such a context anyway.

Citeproc.js (which is basically the reference implementation) seems to just ignore it. We currently created a nested grid. This doesn't translate so well to HTML and is essentially somewhat overcomplicated dead code. In citations, we used to completely disregard any content marked with this display attribute. Now, the attribute is consistently just ignored in places where it is not supported.