Consistently shift links in PDF a bit vertically
This is how it used to work but one place was missed in https://github.com/typst/typst/pull/5420. This PR factors the behaviour out into a function, such that it is not missed again.

This is necessary because e.g. the position of backlinks to footnotes is always at the baseline and if you link directly to it, the text will not be visible since it is right above.