Implement a pre-compile hook for typst compile and watch
This allows the user to automatically execute a script or other command
before compilation starts. While it would be trivial do do this manually
with `./script && typst compile` it becomes extremely useful with `typst
watch`.

The script is executed immediately before compilation starts and
incleded in the typst compilation time output. If the exit status is
non-zero compilation will still run to allow the user to get a partial
document, however a warning is printed.

In order to avoid the case where the script triggers watched files in
`typst watch` mode all watchers are paused during execution. Without
this infinite compilation loops were observed.

Part-Of: #6859