Update link on selector page to point to element functions
Currently the selector docs page of the website mentions element functions, but points to the functions page and not the element functions sub-heading. However, other pages that refer to element functions do point to the element functions section of the function page. This PR updates the link on the selector docs page to point to the element functions sub-section of the functions page.