docs(utils): Rephrase docs for `TypstVersion`
regarding the source of version information. The docs previously
documented what the actual code did *and* the fallback mechanism
implemented in `build.rs`. This was confusing to read since the code
apparently didn't match the docs. In addition, `build.rs` is only a
fallback that can easily be overridden at compile time, which the docs
now reflect.