Add string interpolation
String interpolation is a syntax feature with which values can be
conveniently embedded into strings. It works similar to placing values
in document markup. It can also be though of as syntax sugar for
concatenating the string representation of values into pure string
literals, but the string representation follows slightly different
conversion rules than just `str` or `repr`.

For example, the string representation chosen for an array `a` may be
the same as `repr(a)`, but the string representation of a string `s` is
just `s`.