Spot colors
This commit adds spot colors to precisely request a given pigment in a professional print environment.

I have added the new type `spot-colorant`. Once constructed with a colorant name and a fallback color, the type has a single `.tint(Ratio)` method to construct a new spot color. `spot-colorant` accepts the special values `"all"` (all available colorants) and `none` (no colorants) that are specified in PDF.

For this feature, I had to break out process colors from the Color enum. I have tried my best to dispatch to small, Copyable process types early while keeping the existing API in place.

There is a behavioral change for gradients: They will now eagerly convert to the mixing space instead of doing so during export. This means that errors in conversion (which conversions to spot spaces can yield) will result in an error with a proper span when constructing the gradient instead of an unspanned error or panic during export. This means that `gradient.stops` will now return the spots in the mixing color space instead of their original space. The change is also a speedup since the conversion does not have to be performed for each sample.

Gradients gained the possibility to choose `auto` (new default) as a mixing space. `auto` will result in a spot color if all stops are spot colors using the same colorant and the old default of Oklab otherwise.

For the moment, all exporters will just export the fallback. I have written the necessary code for Krilla to support Separation color spaces. Once that is merged, I will integrate it here and set this to Ready to Review.

Open questions:
- Should the colorant name and fallback color be exposed on `spot-colorant`
- Should the computed fallback color be exposed on spot colors?